Compilation fixes for Windows

+++ mojo/edk/system/channel.cc
@@ -6,6 +6,8 @@
 
 #include <string.h>
 
+#include <algorithm>
+
 #include "base/macros.h"
 #include "base/memory/aligned_memory.h"
 
@@ -30,8 +32,9 @@ Channel::Message::Message(size_t payload_size,
   data_ = static_cast<char*>(base::AlignedAlloc(size_,
                                                 kChannelMessageAlignment));
   Header* header = reinterpret_cast<Header*>(data_);
-  header->num_bytes = size_;
-  header->num_handles = handles_ ? handles_->size() : 0;
+  // TODO: Make sure these casts are safe.
+  header->num_bytes = static_cast<uint32_t>(size_);
+  header->num_handles = static_cast<uint16_t>(handles_ ? handles_->size() : 0);
   header->padding = 0;
 }
 
@@ -40,7 +43,8 @@ Channel::Message::~Message() {
 }
 
 void Channel::Message::SetHandles(ScopedPlatformHandleVectorPtr handles) {
-  header()->num_handles = handles ? handles->size() : 0;
+  // TODO: Make sure this cast is safe.
+  header()->num_handles = static_cast<uint16_t>(handles ? handles->size() : 0);
   std::swap(handles, handles_);
 }
